Final push update for front end
